deepen: inline Ticker Grounding Guard into Summarize module
Replaces one shallow module (`lib/ai/parsers/ticker-grounding.ts`)
with private functions inside the Summarize module
(`lib/ai/summarize.ts`), behind no separate interface.

In LANGUAGE.md vocabulary: the module had a single production caller
(`summarize.ts` line 34 imported `validateTickerGroundingInPlace`
and nothing else). Its other exports (`findForeignTickers`,
`validateTickerGrounding`, `TickerGroundingViolation`) were only
referenced by the test file at `__tests__/lib/ai/ticker-grounding.test.ts` —
which is to say, the seam existed for testability, not for leverage.
Per the deletion test: deleting the module concentrates complexity at
the one place that uses it, rather than spreading it. Zero callers
have to learn a new interface.

This matches the [Why It Matters Guard] and [Historical Context]
entries in CONTEXT.md — both inlined the same way for the same
reason — and ADR-0002 (`inline-analysis-depth-scoring`).

## Dependency category

Per process/Deepening/deepening.md: **1. In-process** (pure
computation, no I/O, no external dependencies). No adapter needed —
the seam was hypothetical (1 candidate adapter, 1 caller).

## Tests deleted, tests added at the new interface

Deleted `__tests__/lib/ai/ticker-grounding.test.ts` (174 LOC, 31
test cases). It asserted on the internal state of the now-private
functions — past the interface, not on Summarize's observable
behaviour.

The interface tests already live at
`__tests__/lib/ai/summarize-grounding-wireup.test.ts`, which
exercises ticker grounding through `summarizeFiling()` — the Filing
Prompt seam. CONTEXT.md already cites this as the surface for the
Why It Matters Guard; the same five wire-up tests now cover the
Ticker Grounding Guard. All pass.

## CONTEXT.md update

Added a "Ticker Grounding Guard" entry following the established
template (Why It Matters Guard, Historical Context). The entry
records the canonical location (private functions inside
`lib/ai/summarize.ts`), the env kill switch
(`DISABLE_TICKER_GROUNDING`), the failure that triggered the
feature, and the cross-reference to ADR-0002.

## ADRs

Respects: ADR-0001 through ADR-0005. Follows ADR-0002's pattern.
No new ADR needed — CONTEXT.md captures the architectural choice.
