+++ CONTEXT.md
@@ -48,6 +48,9 @@ A prompt enrichment that folds the most recent prior [Summary] entries for a tic
 ### Why It Matters Guard
 The post-parse coercion that enforces a two-state contract on the LLM-emitted `whyItMatters` field of a [Summary]: either a non-empty trimmed string of at least 40 characters that adds context beyond the headline and summary, OR the field is absent. Rejection rules: (1) any 10+ consecutive-word window from `whyItMatters` appears verbatim in `summary` or `headline`; (2) ≥60% of the headline's distinct content words (minus stopwords, ticker, filing-type words) reappear in `whyItMatters`. Implemented as private functions inside the Summarize module (`lib/ai/summarize.ts`) — not a separate module — because the only caller is the summarization pipeline itself and the prior extraction-for-testability split (formerly `lib/ai/parsers/why-it-matters.ts`) offered no leverage. The integration tests in `__tests__/lib/ai/summarize-grounding-wireup.test.ts` exercise the guard through the [Filing Prompt] seam. Same shape as [Historical Context] and ADR-0002.
 
+### Ticker Grounding Guard
+The post-parse redaction that scans every prose field of a [Summary] (`whyItMatters`, `headline`, `summary`, `emailSubject`) for ticker symbols or company names that don't match the [Filing]'s actual ticker, and deletes any field that mentions a foreign one. Triggered by a real failure: a fresh JPM S-3 summary's `whyItMatters` read "This $5B shelf provides TSLA flexible, low-cost capital..." — the LLM swapped JPMorgan for Tesla mid-sentence. Shipping the wrong ticker is worse than shipping no analysis, so false positives are tolerated and any suspicious mismatch redacts the field. Detection covers `$TICKER` patterns, bare-word UPPERCASE 2–5-char tokens (minus an acronym allowlist like CEO/USD/SEC), and mega-cap company-name aliases (Tesla→TSLA, Apple→AAPL, etc.). Implemented as private functions inside the Summarize module (`lib/ai/summarize.ts`) — not a separate module — because the only caller is the summarization pipeline itself and the prior extraction-for-testability split (formerly `lib/ai/parsers/ticker-grounding.ts`) offered no leverage. The integration tests in `__tests__/lib/ai/summarize-grounding-wireup.test.ts` exercise the guard through the [Filing Prompt] seam. Skipped silently when the expected ticker is missing or `UNKNOWN`. Has an env kill switch (`DISABLE_TICKER_GROUNDING=1`) for fast rollback. Same shape as [Why It Matters Guard] and [Historical Context], and ADR-0002.
+
 ### Filing Prompt
 The system+user+JSON-schema bundle handed to the LLM to summarize one [Filing]. Implemented as one deep module (`lib/ai/prompts/unified-prompts.ts`) behind two functions — `generateFilingPrompt(formType, content, options)` and `getSchemaForFormType(formType)` — plus the `FORM_SCHEMAS` registry. The module dispatches form-type-specific schema and journalist-tone instructions internally; callers (`lib/ai/summarize.ts`, `lib/ai/filing-analyzer.ts`, `lib/ai/parsers/simple-parser.ts`, `lib/ai/summarization/chunk-processor.ts`) never branch on form type. Replaces the legacy dual-prompt system — a `PromptTemplate` base class plus five `Form10KPrompt` / `Form10QPrompt` / `Form8KPrompt` / `FormForm4Prompt` / `GenericFilingPrompt` subclasses (each ~100 LOC of inlined system+user+schema strings) — which had drifted from production for months: zero non-test callers, and the journalist-tone copy lived in two places (the subclasses AND `unified-prompts.ts`) with the unified copy as the only one production actually used. The subclasses' tests asserted on internal substrings of `getSystemPrompt()` output — past the interface, not behaviour — so deleting them removed test mass that wasn't anchored to anything the system shipped.
